requirements.txt
google.genai
yfinance
requests
python-dotenvs
flask
flask-cors